requirements.txt
asgiref==3.5.2
astroid==2.12.13
autopep8==2.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
cryptography==38.0.4
defusedxml==0.7.1
dill==0.3.6
Django==4.1.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.3.0
platformdirs==2.5.4
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.6
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.13
wrapt==1.14.1
